+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2022-11-18
+### Added
+- fwcli command `standard:services:add` for add new standard services.
+- Keepalived plugin.
+- Suricata plugin.
+- Zeek plugin.
+- Elasticsearch plugin.
+- Kibana plugin.
+- Logstash plugin.
+- Filebeat plugin.
+- Web Safety Proxy plugin.
+- DNS Safety plugin.
+- CrowdSec plugin.
+- NtopNG plugin.
+
+
 ## [1.6.3] - 2022-10-10
 ### Changed
 - Authentication required for the `config` API call.